Huge initial commit to the repository; call it v0.X.
Been developing these for some time without using git, just
using eclipse's inbuilt version control; making this commit
to allow committing the code to github.  Mostly functional;
a few tweaks to add here and there.

1.)  DozcalDroid crashes when adding intermittent event in
every month of the year; however, the event is successfully
added.
2.)  Should add ability to select arbitrary characters for X
and E; as it is, you can enter them as * or # or X or E or a
or b, but they all output as X and E.
3.)  Others.